+++ NEWS.md
@@ -27,6 +27,10 @@
 * `check_outliers()` now also works for meta-analysis models from packages
   *metafor* and *meta*.
 
+* `plot()` for `performance::check_model()` no longer produces a normal QQ plot
+  for GLMs. Instead, it now shows a half-normal QQ plot of the absolute value
+  of the standardized deviance residuals.
+
 ## Bug fixes
 
 * Fixed issue in `print()` method for `check_collinearity()`, which could mix